horarios

+++ backend/horas/src/main/java/meson/horas/services/EmpleadoService.java
@@ -2,13 +2,15 @@ package meson.horas.services;
 
 import jakarta.transaction.Transactional;
 import meson.horas.entities.Empleado;
+import meson.horas.entities.Horario;
 import meson.horas.repositories.EmpleadoRepository;
+import meson.horas.repositories.HorarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.HttpClientErrorException;
 
-import java.net.http.HttpConnectTimeoutException;
+import java.util.HashMap;
 import java.util.List;
 
 @Service
@@ -18,8 +20,22 @@ public class EmpleadoService {
     @Autowired
     EmpleadoRepository empleadoRepository;
 
+    @Autowired
+    HorarioRepository horarioRepository;
+
     public List<Empleado> getEmpleados() {
-        return empleadoRepository.findAll();
+
+        var empleados = empleadoRepository.findAll();
+
+        empleados.forEach(empleado -> {
+            var horarios = new HashMap<String, Horario>();
+            horarioRepository.findByEmpleado(empleado).forEach(horario -> {
+                horarios.put(horario.getDiaSemana().toString(), horario);
+            });
+            empleado.setHorarios(horarios);
+        });
+
+        return empleados;
     }
 
     public Empleado getEmpleado(Long id) {
@@ -38,13 +54,32 @@ public class EmpleadoService {
     public Empleado editEmpleado(Empleado empleado) {
 
         if(empleado.getId() == null) {
-            throw new HttpClientErrorException(HttpStatus.BAD_REQUEST);
+            throw new HttpClientErrorException(HttpStatus.BAD_REQUEST, "El id del empleado no puede ser nulo");
         }
+
+        var horarios = empleado.getHorarios();
+
+        try{
+            if(horarios != null) {
+                horarios.forEach((k, v) -> {
+                    v.setEmpleado(empleado);
+                    var horario = horarioRepository.findByEmpleadoAndDiaSemana(empleado, v.getDiaSemana());
+                    v.setId(horario == null ? null : horario.getId());
+                    horarioRepository.save(v);
+                });
+            }
+        }catch(Exception e){
+            throw new HttpClientErrorException(HttpStatus.BAD_REQUEST, "Error en formato de los horarios");
+        }
+
         return empleadoRepository.save(empleado);
 
     }
 
     public void deleteEmpleado(Long id) {
+        if(!empleadoRepository.existsById(id)) {
+            throw new HttpClientErrorException(HttpStatus.NOT_FOUND, "El empleado no existe");
+        }
         empleadoRepository.deleteById(id);
     }
